Add support for unicode filename in zip file
Since there is no standard what kind of encoding the filename in the ZIP file have,
and so far we assumed 'utf-8', it failed if the filename was unicode.

I was not able to reproduce this on OSX, since the filenames turned out always
as utf-8.

But you can reproduce it by constructing your own ZIP file using a unicode filename

>>> zip_ = zipfile.ZipFile(path+'/unicode.zip', 'w')
>>> zip_.write(path_to_file, u'äöü/test.txt')
>>> zip_.close()